Added helper Completing fcm

+++ src/migrations/2015_01_01_100000_create_fcm_projects_table.php
@@ -14,10 +14,10 @@ class CreateFcmProjectsTable extends Migration
     public function up()
     {
         Schema::create('fcm_projects', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('name',191)->nullable(false)->comment('project name in the firebase.io');
             $table->string('server_key',191)->nullable(false)->comment('server key in the firebase.io');
-            $table->unsignedInteger('user_id')->nullable(false)->comment('relation to users table');
+            $table->unsignedBigInteger('user_id')->nullable(false)->comment('relation to users table');
             $table->timestamps();
 
             $table->foreign('user_id')->references('id')->on('users');
